Make hasMetaChanges selector private
